Security Update
Basic Authentication

+++ src/main/java/webService/MenusServices.java
@@ -1,5 +1,8 @@
 package webService;
 
+import java.util.Arrays;
+import java.util.HashSet;
+
 import javax.ejb.EJB;
 import javax.ejb.Stateless;
 import javax.ws.rs.DELETE;
@@ -9,7 +12,10 @@ import javax.ws.rs.PUT;
 import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
+import javax.ws.rs.core.Context;
+import javax.ws.rs.core.HttpHeaders;
 import javax.ws.rs.core.MediaType;
+import javax.ws.rs.core.Response;
 
 import database.entity.Menu;
 import database.entity.MenuDAO;
@@ -20,29 +26,41 @@ import database.entity.MenuDAO;
 public class MenusServices {
 	@EJB
 	private MenuDAO menuDao;
-
+	@EJB
+	private RequestInterceptor interceptor;
 	@POST
 	@Path("/createMenu")
-	public Menu createNewMenu(Menu m) {
-		return menuDao.createMenu(m);
+	public Response createNewMenu(@Context HttpHeaders hHeaders,Menu m) {
+		if(interceptor.process(new HashSet<String>(Arrays.asList(new String[]{"admin"})), hHeaders)){
+			return Response.status(200).entity(menuDao.createMenu(m)).build();
+		}
+		return Response.status(401).entity("Unauthorized").build();
 	}
-
 	@PUT
 	@Path("/updateMenu")
-	public Menu updateMenu(Menu menu) {
+	public Response updateMenu(@Context HttpHeaders hHeaders,Menu menu) {
 		//the new menu object should contain reference ID
-		return menuDao.updateMenu(menu);
+		if(interceptor.process(new HashSet<String>(Arrays.asList(new String[]{"admin"})), hHeaders)){
+			return Response.status(200).entity(menuDao.updateMenu(menu)).build();
+		}
+		return Response.status(401).entity("Unauthorized").build();
 	}
 
 	@GET
 	@Path("/{menuid}")
-	public Menu getMenuById(@PathParam("menuid") int menuid) {
-		return menuDao.getMenuById(menuid);
+	public Response getMenuById(@PathParam("menuid") int menuid) {
+		return Response.status(200).entity(menuDao.getMenuById(menuid)).build();
 
 	}
+
 	@DELETE
 	@Path("/delete/{menuid}")
-	public void deleteMenu(@PathParam("menuid") int menuid){
-		menuDao.deleteMenu(menuid);
+	public Response deleteMenu(@Context HttpHeaders hHeaders,@PathParam("menuid") int menuid){
+		if(interceptor.process(new HashSet<String>(Arrays.asList(new String[]{"admin"})), hHeaders)){
+			menuDao.deleteMenu(menuid);
+			return Response.status(200).entity("Deleted").build();
+		}
+		return Response.status(401).entity("Unauthorized").build();
+		
 	}
 }
