+++ src/main/java/webService/OrderService.java
@@ -1,7 +1,10 @@
 package webService;
 
+import java.util.Arrays;
+import java.util.HashSet;
 import java.util.List;
 
+import javax.annotation.security.RolesAllowed;
 import javax.ejb.EJB;
 import javax.ejb.Stateless;
 import javax.ws.rs.Consumes;
@@ -11,7 +14,10 @@ import javax.ws.rs.POST;
 import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
+import javax.ws.rs.core.Context;
+import javax.ws.rs.core.HttpHeaders;
 import javax.ws.rs.core.MediaType;
+import javax.ws.rs.core.Response;
 
 import database.entity.Orders;
 import database.entity.OrdersDAO;
@@ -19,9 +25,12 @@ import database.entity.OrdersDAO;
 @Path("/order")
 @Stateless
 @Produces(MediaType.APPLICATION_JSON)
+@RolesAllowed({"customer","admin"})
 public class OrderService {
 	@EJB
 	private OrdersDAO ordersDao;
+	@EJB
+	private RequestInterceptor interceptor;
 	
 	@POST
 	@Path("/createOrder")
@@ -41,7 +50,12 @@ public class OrderService {
 	}
 	@DELETE
 	@Path("/delete/{orderid}")
-	public void deleteOrder(@PathParam("orderid")int orderid ){
-		ordersDao.removeOrder(orderid);
+	public Response deleteOrder(@Context HttpHeaders hHeaders,@PathParam("orderid")int orderid ){
+		if(interceptor.process(new HashSet<String>(Arrays.asList(new String[]{"admin","customer"})), hHeaders)){
+			ordersDao.removeOrder(orderid);
+			return Response.status(200).entity("removed").build();
+		}
+		return Response.status(401).entity("Unauthorized").build();
+		
 	}
 }
