+++ src/main/java/webService/RestaurantService.java
@@ -1,9 +1,10 @@
 package webService;
 
+import java.util.Arrays;
+import java.util.HashSet;
 import java.util.List;
 
 import javax.ejb.EJB;
-import javax.ejb.LocalBean;
 import javax.ejb.Stateless;
 import javax.ws.rs.Consumes;
 import javax.ws.rs.DELETE;
@@ -13,11 +14,12 @@ import javax.ws.rs.PUT;
 import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
+import javax.ws.rs.core.Context;
+import javax.ws.rs.core.HttpHeaders;
 import javax.ws.rs.core.MediaType;
+import javax.ws.rs.core.Response;
 
-import database.entity.Manager;
 import database.entity.Menu;
-import database.entity.Orders;
 import database.entity.Restaurant;
 import database.entity.RestaurantDAO;
 import database.entity.RestaurantLocation;
@@ -25,62 +27,90 @@ import database.entity.RestaurantLocation;
 @Path("/restaurants")
 @Stateless
 @Produces(MediaType.APPLICATION_JSON)
+
 public class RestaurantService {
 	@EJB
 	private RestaurantDAO restaurantDao;
+	@EJB
+	private RequestInterceptor interceptor;
 	
 	@POST
 	@Path("/createRestaurant")
 	@Consumes(MediaType.APPLICATION_JSON)
-	public Restaurant createRestaurant(Restaurant r){
-		
-		return restaurantDao.createRestaurant(r);
+	public Response createRestaurant(@Context HttpHeaders hHeaders,Restaurant r) {
+		if(interceptor.process(new HashSet<String>(Arrays.asList(new String[]{"admin"})), hHeaders)){
+			return Response.status(200).entity(restaurantDao.createRestaurant(r)).build();
+		}
+		return Response.status(401).entity("Unauthorized").build();
 	}
-	
 	@PUT
 	@Path("/updateRestaurant")
-	public Restaurant updateRestaurantInfo(Restaurant r){
-		return restaurantDao.update(r);
+	public Response updateRestaurantInfo(@Context HttpHeaders hHeaders,Restaurant r) {
+		if(interceptor.process(new HashSet<String>(Arrays.asList(new String[]{"admin"})), hHeaders)){
+			return Response.status(200).entity( restaurantDao.update(r)).build();
+		}
+		return Response.status(401).entity("Unauthorized").build();
 	}
+
 	@GET
 	@Path("/{id}")
-	public Restaurant getRestaurantById(@PathParam("id")int id){
+	public Restaurant getRestaurantById(@PathParam("id") int id) {
 		return restaurantDao.getRestaurantById(id);
-		
+
 	}
+
 	@GET
 	@Path("/{id}/orders")
-	public List<Orders> getRestaurantOrdersById(@PathParam("id")int id){
-		return restaurantDao.getRestaurantOrdersById(id);
-		
+	public Response getRestaurantOrdersById(@Context HttpHeaders hHeaders,@PathParam("id") int id) {
+		if(interceptor.process(new HashSet<String>(Arrays.asList(new String[]{"admin"})), hHeaders)){
+			return Response.status(200).entity(restaurantDao.getRestaurantOrdersById(id)).build();
+		}
+		return Response.status(401).entity("Unauthorized").build();
+
 	}
+
 	@GET
 	@Path("/{id}/menus")
-	public List<Menu> getRestaurantMenusById(@PathParam("id")int id){
+	public List<Menu> getRestaurantMenusById(@PathParam("id") int id) {
 		return restaurantDao.getRestaurantMenusById(id);
-		
+
 	}
+
 	@GET
 	@Path("/restaurant/locations/{locationid}")
-	public RestaurantLocation getBranchRestaurantInfoById(@PathParam("locationid")int locationid){
+	public RestaurantLocation getBranchRestaurantInfoById(
+			@PathParam("locationid") int locationid) {
 		return restaurantDao.getBranchRestaurantInfoById(locationid);
-		
+
 	}
+
 	@GET
 	@Path("/allrestaurant/locations")
-	public List<Restaurant> getAllRestaurantLocations(){
+	public List<Restaurant> getAllRestaurantLocations() {
 		return restaurantDao.getAllRestaurantLocations();
-		
+
 	}
 	@GET
+	@Path("/restaurantByLocationId/{locationid}")
+	public Restaurant getRestaurantByLocationId(@PathParam("locationid") int locationid){
+		return restaurantDao.getRestaurantByLocationId(locationid);
+	}
+
+	@GET
 	@Path("/{id}/locations")
-	public List<RestaurantLocation> getBranchRestaurantsById(@PathParam("id")int id){
+	public List<RestaurantLocation> getBranchRestaurantsById(
+			@PathParam("id") int id) {
 		return restaurantDao.getBranchRestaurantsById(id);
-		
+
 	}
 	@DELETE
 	@Path("/delete/{restaurantid}")
-	public void deleteARestaurant(@PathParam("restaurantid")int restaurantid){
-		restaurantDao.deleteARestaurant(restaurantid);
+	public Response deleteARestaurant(@Context HttpHeaders hHeaders,@PathParam("restaurantid") int restaurantid) {
+		if(interceptor.process(new HashSet<String>(Arrays.asList(new String[]{"admin"})), hHeaders)){
+			restaurantDao.deleteARestaurant(restaurantid);
+			return Response.status(200).entity("Deleted").build();
+		}
+		return Response.status(401).entity("Unauthorized").build();
+		
 	}
 }
